Allow the unit tests to run in parallel
