--- .mocharc.json
@@ -1,3 +1,0 @@
-{
-  "timeout": 20000
-}
