+++ tests/tokens/periphery/ethdkg/complete-round.test.ts
@@ -1,0 +1,13 @@
+import { validators4 } from "./assets/4-validators-successful-case";
+import { validators10 } from "./assets/10-validators-successful-case";
+import { completeETHDKGRound } from "./setup";
+
+describe("Happy Path", () => {
+  it("completes happy path with 4 validators", async function () {
+    await completeETHDKGRound(validators4);
+  });
+
+  it("completes happy path with 10 validators", async function () {
+    await completeETHDKGRound(validators10);
+  });
+});
