+++ tests/tokens/periphery/ethdkg/phases/03-dispute-share-distribution.test.ts
@@ -1,0 +1,11 @@
+import { validators4 } from "../assets/4-validators-successful-case";
+import { ethers } from "hardhat";
+import {
+  endCurrentPhase,
+  endCurrentAccusationPhase,
+  distributeValidatorsShares,
+  startAtDistributeShares,
+  expect,
+} from "../setup";
+
+describe("Distribute bad shares accusation", () => {});
