+++ tests/tokens/periphery/ethdkg/phases/07-dispute-gpkj-submission.test.ts
@@ -1,0 +1,10 @@
+import { validators4 } from "../assets/4-validators-successful-case";
+import { ethers } from "hardhat";
+import {
+  endCurrentPhase,
+  endCurrentAccusationPhase,
+  distributeValidatorsShares,
+  startAtDistributeShares,
+  expect,
+} from "../setup";
+
