+++ tests/tokens/periphery/ethdkg/phases/08-completion.test.ts
@@ -1,0 +1,2 @@
+import { validators4 } from "../assets/4-validators-successful-case";
+
